+++ packages/placeholders/CHANGELOG.md
@@ -1,3 +1,10 @@
+# [@solid-design-system/placeholders-v1.8.0](https://github.com/solid-design-system/solid/compare/placeholders/1.7.1...placeholders/1.8.0) (2024-11-08)
+
+
+### Features
+
+* create docs package for solid components ([#1550](https://github.com/solid-design-system/solid/issues/1550)) ([7555d18](https://github.com/solid-design-system/solid/commit/7555d182abfa34a23521c839180ae4b67604717e)), closes [#1131](https://github.com/solid-design-system/solid/issues/1131)
+
 # [@solid-design-system/placeholders-v1.7.1](https://github.com/solid-design-system/solid/compare/placeholders/1.7.0...placeholders/1.7.1) (2024-10-16)